+++ inst/include/mlpack/core/optimizers/sgd/test_function.hpp
@@ -39,17 +39,17 @@ class SGDTestFunction
   SGDTestFunction() { }
 
   //! Return 3 (the number of functions).
-  size_t NumFunctions() const { return 3; }
+  long NumFunctions() const { return 3; }
 
   //! Get the starting point.
   arma::mat GetInitialPoint() const { return arma::mat("6; -45.6; 6.2"); }
 
   //! Evaluate a function.
-  double Evaluate(const arma::mat& coordinates, const size_t i) const;
+  double Evaluate(const arma::mat& coordinates, const long i) const;
 
   //! Evaluate the gradient of a function.
   void Gradient(const arma::mat& coordinates,
-                const size_t i,
+                const long i,
                 arma::mat& gradient) const;
 };
